fix delete byId

+++ packages/domain/lib/usecase/get_movie_list_usecase.dart
@@ -19,7 +19,6 @@ class GetMovieListUseCase
   @override
   Future<List<MovieCache>> call(MovieType currentMovieType) async {
     final List<MovieCache> jsonMovies = [];
-    final List<MovieCache> jsonMoviesFromCache = [];
     final List<int> movieCacheIds = [];
     if (await _isTimestampNewDay(
       DateTime.now(),
@@ -43,19 +42,25 @@ class GetMovieListUseCase
       final List<int> idsToDeleteFromDB = movieCacheIds
           .where((e1) => jsonMovies.where((e) => e.trakt == e1).isEmpty)
           .toList();
-      await _localStorageRepository.deleteByIds(
-          idsToDeleteFromDB, currentMovieType);
+      if (idsToDeleteFromDB.isNotEmpty) {
+        await _localStorageRepository.deleteByIds(
+          idsToDeleteFromDB,
+          currentMovieType,
+        );
+      }
       final List<MovieCache> moviesToAddToDB = jsonMovies
           .where(
               (movie) => movieCacheIds.where((id) => movie.trakt == id).isEmpty)
           .toList();
-      await _localStorageRepository.saveMoviesToCache(
-          moviesToAddToDB, currentMovieType);
+      if (moviesToAddToDB.isNotEmpty) {
+        await _localStorageRepository.saveMoviesToCache(
+          moviesToAddToDB,
+          currentMovieType,
+        );
+      }
       return jsonMovies;
     } else {
-      jsonMoviesFromCache.addAll(
-          await _localStorageRepository.getMoviesFromCache(currentMovieType));
-      return jsonMoviesFromCache;
+      return await _localStorageRepository.getMoviesFromCache(currentMovieType);
     }
   }
